update curriculum to langgraph based
