+++ config.py
@@ -7,11 +7,9 @@ load_dotenv()
 
 class Config:
     # LLM 설정
-    LLM_BASE_URL = "https://api.friendli.ai/serverless/v1"
+    LLM_BASE_URL = "https://api.friendli.ai/dedicated/v1"
     LLM_API_KEY = os.getenv("LLM_API_KEY")
-    LLM_MODEL = "K-intelligence/Midm-2.0-Base-Instruct"
-    # LLM_MODEL = "midm-2.0-base-bf16"
-    
+    LLM_MODEL = "depnct19r37qy14"    
     
     LLM_TEMPERATURE = 0.0
     LLM_MAX_TOKENS = 2048  # 응답 생성용 최대 토큰 수
