--- data/backups/test_backup.json
@@ -1,249 +1,0 @@
-{
-  "backup_timestamp": "2025-09-04T21:27:33.696926",
-  "curriculums": {
-    "persistent_user": [
-      {
-        "title": "React Native Mobile Development Learning Path",
-        "level": "intermediate",
-        "duration_weeks": 4,
-        "modules": [
-          {
-            "week": 1,
-            "title": "Advanced Concepts",
-            "objectives": [
-              "Understand key concepts of advanced",
-              "Apply React Native Mobile Development in advanced context",
-              "Complete practical exercises"
-            ],
-            "resources": {
-              "primary": [],
-              "supplementary": [],
-              "practice": []
-            },
-            "estimated_hours": 10
-          },
-          {
-            "week": 2,
-            "title": "Best Practices",
-            "objectives": [
-              "Understand key concepts of patterns",
-              "Apply React Native Mobile Development in patterns context",
-              "Complete practical exercises"
-            ],
-            "resources": {
-              "primary": [],
-              "supplementary": [],
-              "practice": []
-            },
-            "estimated_hours": 12
-          },
-          {
-            "week": 3,
-            "title": "Real-world Applications",
-            "objectives": [
-              "Understand key concepts of applications",
-              "Apply React Native Mobile Development in applications context",
-              "Complete practical exercises"
-            ],
-            "resources": {
-              "primary": [],
-              "supplementary": [],
-              "practice": []
-            },
-            "estimated_hours": 14
-          },
-          {
-            "week": 4,
-            "title": "Complex Projects",
-            "objectives": [
-              "Understand key concepts of projects",
-              "Apply React Native Mobile Development in projects context",
-              "Complete practical exercises"
-            ],
-            "resources": {
-              "primary": [],
-              "supplementary": [],
-              "practice": []
-            },
-            "estimated_hours": 16
-          }
-        ],
-        "recommended_resources": {
-          "courses": [],
-          "documentation": [],
-          "tools": []
-        },
-        "generation_method": "rule_based",
-        "user_id": "persistent_user",
-        "focus_areas": [
-          "navigation",
-          "state management"
-        ],
-        "generated_at": "2025-09-04T21:27:33.692489",
-        "id": 0,
-        "created_at": "2025-09-04T21:27:33.692503",
-        "curriculum_id": 0
-      }
-    ]
-  },
-  "daily_plans": {
-    "persistent_user_0": [
-      {
-        "week_number": 1,
-        "module_title": "Advanced Concepts",
-        "daily_plans": [
-          {
-            "day": 1,
-            "date": "2025-09-04",
-            "type": "concept",
-            "focus": "Learning new concepts",
-            "duration_hours": 2.6,
-            "tasks": [
-              {
-                "task": "Read primary resources",
-                "duration": "45 min",
-                "completed": true
-              },
-              {
-                "task": "Take notes on key concepts",
-                "duration": "30 min",
-                "completed": false
-              },
-              {
-                "task": "Watch explanatory videos",
-                "duration": "30 min",
-                "completed": false
-              }
-            ],
-            "resources_for_today": {
-              "primary": [],
-              "practice": []
-            }
-          },
-          {
-            "day": 2,
-            "date": "2025-09-05",
-            "type": "practice",
-            "focus": "Hands-on practice",
-            "duration_hours": 2.8,
-            "tasks": [
-              {
-                "task": "Complete coding exercises",
-                "duration": "60 min",
-                "completed": false
-              },
-              {
-                "task": "Solve practice problems",
-                "duration": "45 min",
-                "completed": false
-              },
-              {
-                "task": "Debug and refactor code",
-                "duration": "30 min",
-                "completed": false
-              }
-            ],
-            "resources_for_today": {
-              "primary": [],
-              "practice": []
-            }
-          },
-          {
-            "day": 3,
-            "date": "2025-09-06",
-            "type": "review",
-            "focus": "Review and consolidation",
-            "duration_hours": 2.4,
-            "tasks": [
-              {
-                "task": "Review notes from the week",
-                "duration": "30 min",
-                "completed": false
-              },
-              {
-                "task": "Summarize key learnings",
-                "duration": "30 min",
-                "completed": false
-              },
-              {
-                "task": "Identify knowledge gaps",
-                "duration": "20 min",
-                "completed": false
-              }
-            ],
-            "resources_for_today": {
-              "primary": [],
-              "practice": []
-            }
-          },
-          {
-            "day": 4,
-            "date": "2025-09-07",
-            "type": "project",
-            "focus": "Project work",
-            "duration_hours": 2.2,
-            "tasks": [
-              {
-                "task": "Work on module project",
-                "duration": "90 min",
-                "completed": false
-              },
-              {
-                "task": "Document progress",
-                "duration": "20 min",
-                "completed": false
-              },
-              {
-                "task": "Seek feedback",
-                "duration": "15 min",
-                "completed": false
-              }
-            ],
-            "resources_for_today": {
-              "primary": [],
-              "practice": []
-            }
-          },
-          {
-            "day": 5,
-            "date": "2025-09-08",
-            "type": "concept",
-            "focus": "Learning new concepts",
-            "duration_hours": 2.6,
-            "tasks": [
-              {
-                "task": "Read primary resources",
-                "duration": "45 min",
-                "completed": false
-              },
-              {
-                "task": "Take notes on key concepts",
-                "duration": "30 min",
-                "completed": false
-              },
-              {
-                "task": "Watch explanatory videos",
-                "duration": "30 min",
-                "completed": false
-              }
-            ],
-            "resources_for_today": {
-              "primary": [],
-              "practice": []
-            }
-          }
-        ],
-        "created_at": "2025-09-04T21:27:33.694052"
-      }
-    ]
-  },
-  "learning_progress": {
-    "persistent_user_0_1": {
-      "week_number": 1,
-      "day_number": 1,
-      "task_completed": true,
-      "week_progress_percentage": 6.67,
-      "updated_at": "2025-09-04T21:27:33.695458"
-    }
-  }
-}
