+++ pyproject.toml
@@ -18,4 +18,9 @@ dependencies = [
     "pydantic>=2.11.7",
     "python-dotenv>=1.1.1",
     "sentence-transformers>=5.1.0",
+    "fastmcp>=0.3.0",
+    "httpx>=0.25.2",
+    "uvicorn>=0.24.0",
+    "aiofiles>=23.2.1",
+    "python-multipart>=0.0.6",
 ]
