+++ requirements.txt
@@ -13,6 +13,10 @@ langgraph>=0.0.30
 mcp>=1.13.1
 langchain-mcp-adapters>=0.1.0
 
+# 추가 의존성 (LangGraph 시스템용)
+beautifulsoup4>=4.12.0
+python-dotenv>=1.0.0
+
 # 기존 프로젝트 의존성
 fastmcp>=0.3.0
